requirements.txt
Flask==3.0.3
itsdangerous==2.2.0
Werkzeug==3.0.3
Jinja2==3.1.4
python-dotenv==1.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.26.4
